requirements.txt
pandas
numpy
scikit-learn
xgboost
SQLAlchemy
asyncpg
websockets
httpx
web3
python-dotenv
ccxt
binance-futures-connector
pybit
psycopg2-binary
alembic
tardis-client
telethon
pyrogram
